Create an `interactions` module; Move <Button /> over
