+++ src/imports/components/Sections/ClientFooter/index.js
@@ -1,7 +1,7 @@
 import React, {PureComponent} from 'react';
 import PropTypes from 'prop-types';
 import {RightArrowIcon} from 'icons';
-import Button from 'components/Button';
+import Button from 'components/interactions/Button';
 
 import './index.styl';
 import './mobile/index.styl';
