+++ src/pages/Landing/sections/Header/index.js
@@ -6,7 +6,7 @@ import {trackEvent} from 'analytics';
 import NavigationBar from 'components/NavigationBar';
 import Heading from 'components/typography/Heading';
 import Paragraph from 'components/typography/Paragraph';
-import Button from 'components/Button';
+import Button from 'components/interactions/Button';
 
 import Illustration from './components/Illustration';
 import {ExploreButtonIcon} from './icons';
