+++ src/pages/Page404/index.js
@@ -4,7 +4,7 @@ import Loader from 'components/Loader';
 import NavigationBar from 'components/NavigationBar';
 import Heading from 'components/typography/Heading';
 import Paragraph from 'components/typography/Paragraph';
-import Button from 'components/Button';
+import Button from 'components/interactions/Button';
 import {darkBlue, blue} from 'styles/config';
 
 import {LetsGoHomeButtonIcon} from './icons';
